pyproject.toml
[project]
name = "podman-mcp-server"
version = "0.1.0"
description = "Podman MCP Server generated from OpenAPI spec"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "fastmcp>=2.3.0rc1",
  "httpx<1.0.0",
  "pyyaml>=6.0.0",
  "prance>=23.6.21.0",
  "openapi-spec-validator>=0.7.0",
  "anyio>=4.13.0",
  "uvicorn>=0.44.0",
  "starlette>=1.0.0",
  "redis>=7.4.0",
  "py-key-value-aio[redis]>=0.4.4",
]

[dependency-groups]
dev = ["pytest>=9.0.3", "ruff>=0.15.10"]
